Rename lnd_file to LNDFile
