+++ src/Gui.cpp
@@ -28,13 +28,13 @@
 #include "MeshViewer.h"
 #include "Profiler.h"
 
+#include <LNDFile.h>
 #include <bgfx/bgfx.h>
 #include <bgfx/embedded_shader.h>
 #include <bx/math.h>
 #include <bx/timer.h>
 #include <imgui.h>
 #include <imgui_widget_flamegraph.h>
-#include <lnd_file.h>
 #ifdef _WIN32
 #include <SDL2/SDL_syswm.h>
 #endif
